test/requirements.txt
pillow
pyyaml
